:construction: update code

+++ docker-entrypoint.d/00-nginx-entrypoint.sh
@@ -9,21 +9,27 @@ NGX_CONF_DIR="/etc/nginx/conf.d"
 NGX_WORKER_PROC=${NGX_WORKER_PROCESSES:-"1"}
 
 _main() {
-  ### nginx.conf
-  if [ -f $NGX_CONF_FILE ]; then
-      rm -f $NGX_CONF_FILE
-  fi
-
-  cp -f $NGX_CONF_TPL_FILE $NGX_CONF_FILE
-
-  sed -i "s/NGX_WORKER_PROC/$NGX_WORKER_PROC/g" $NGX_CONF_FILE
-
-  # default.conf
-  if [ -f $WORKSPACE/deploy/nginx/default.conf ]; then
-      if [ -f $NGX_CONF_DIR/default.conf ]; then
-        rm -f $NGX_CONF_DIR/default.conf
+  if [ -f /usr/sbin/nginx ]; then
+      ### nginx.conf
+      if [ -f $NGX_CONF_FILE ]; then
+        rm -f $NGX_CONF_FILE
       fi
-      cp -f $WORKSPACE/deploy/nginx/default.conf $NGX_CONF_DIR/default.conf
+
+      if [ -f $NGX_CONF_TPL_FILE ]; then
+        cp -f $NGX_CONF_TPL_FILE $NGX_CONF_FILE
+      fi
+
+      if [ -f $NGX_CONF_FILE ]; then
+        sed -i "s/NGX_WORKER_PROC/$NGX_WORKER_PROC/g" $NGX_CONF_FILE
+      fi
+
+      # default.conf
+      if [ -f $WORKSPACE/deploy/nginx/default.conf ]; then
+          if [ -f $NGX_CONF_DIR/default.conf ]; then
+            rm -f $NGX_CONF_DIR/default.conf
+          fi
+          cp -f $WORKSPACE/deploy/nginx/default.conf $NGX_CONF_DIR/default.conf
+      fi
   fi
 
 }
